`Add BiMap class and refactor LocalizationManager`
Added a new `BiMap` class to `bimap.py` for bidirectional mapping and imported it into `__init__.py` and `localization_manager.py`. The `LocalizationManager` class was significantly refactored, including the addition of `fallback_language` and `lazy_load` attributes, renaming of `default_language` to `active_language`, and replacement of multiple attributes with `dictionaries` storing `BiMap` instances. Several methods were added for language management, dictionary processing, and observer notification. The `translate` method was renamed to `localize` and a `reverse_localize` method was added. The default logging level in `utils.py` was changed from `INFO` to `WARNING`.