Refactor `localize_with_params` in `LocalizationManager`
The `localize_with_params` method in the `LocalizationManager` class has been significantly refactored. The method now accepts a variable number of positional arguments instead of keyword arguments. The docstring has been updated to reflect this change. The method now uses the `format` method of the localized string to substitute the arguments into the string, and this is done inside a `try` block to catch `IndexError` exceptions. If an `IndexError` is caught, an error message is logged and the untranslated template string is returned.