Refactor logging module and add tests
This commit includes a major refactor of the logging module. The `Logger` class has been updated to use the new `LoggerConfig` and `_LoggerCore` classes from the `logging/internal/_logger_core.py` file. The `_logger_core.py` and `logger_mixin.py` files were removed, and their functionality was moved to the `logging/internal` directory. The `LoggerConfig` now includes options for enabling/disabling console and file logging, and the `_LoggerCore` now includes separate methods for logging to the console and to a file. The `Logger` class methods for adding, changing, and removing logger configurations were removed, and methods for adding and retrieving loggers were added. The logging methods were also updated to use the new `_LoggerCore` methods. The `__init__.py` file in the `logging` directory was modified to reflect these changes. A new `test_logger.py` file was added to test the functionality of the `Logger` class.