Updated `create_tree` method in `FileTreeView` class
The `create_tree` method in the `FileTreeView` class has been updated to take an additional argument `root`. This argument is now passed when the `create_tree` method is called in the `__init__` method of the same class. Previously, the `create_tree` method was called with only `expand_root_node` as an argument. Now, it is called with both `root` and `expand_root_node` as arguments.