Add CustomCTKButton class and update CustomTooltip class
Added a new class `CustomCTKButton` in `custom_ctk_button.py` that extends `CTkButton` with additional methods and properties. This class includes parameters for button and tooltip properties, and methods for updating localization, setting text, creating tooltips, and setting tooltip text and delay. Also, added a new `update_localization` method that updates the button and tooltip text using the localization function. The `set_text` method is used to set the button and optionally the tooltip text. The `create_tooltip` method creates a tooltip with given text and properties. The `set_tooltip_text` and `set_tooltip_delay` methods are used to set the tooltip text and delay respectively. In the `CustomTooltip` class in `custom_tooltip.py`, a new `set_text` method was added to set the tooltip text.