Added `get_locale` method to `Locales` class
This commit introduces a new class method `get_locale` in the `Locales` class. The method accepts a `locale_code` and returns the corresponding `Locale` object. It achieves this by iterating over all locales from the `get_locales` method and checking for a match with the `locale` attribute of each `Locale` object. If a match is found, the `Locale` object is returned.