Refactor localization in GUI classes with Localizer and LocalizationKey
This commit includes a major refactoring of the localization system in the GUI classes. The `typing.Callable` import was removed and replaced with `typing.Union` to allow text parameters to accept either a string or a `LocalizationKey`. The `Localizer` and `LocalizationKey` classes were imported from the `GuiFramework.utilities.localization` package for text localization in GUI elements. The `loc_func` parameter was removed from the `__init__` methods of the `CustomCTKButton`, `CustomCTKCheckbox`, and `CustomCTKLabel` classes and replaced with `Localizer` class methods. The `update_localization` methods in these classes now subscribe to the `Localizer.EVENT_LANGUAGE_CHANGED` event to update GUI text when the language changes. The `set_text`, `create_tooltip`, and `set_tooltip_text` methods now use the `Localizer.get_localized_string` method for text localization. Lastly, the `update` method call was commented out in the `CustomCTKOptionMenu` class's `update_localization` method.