Updated LocalizationKey and LocalizationKeyGenerator classes
In `localization_key.py`, the `uuid` module was imported and the `__init__` method of the `LocalizationKey` class was modified to accept an optional `key` argument. If no `key` is provided, a new UUID is generated and used as the key. The `get_localization_key_for_string` method was removed from the `LocalizationKey` class.

In `localization_key_generator.py`, the import statement for the `black` module and the call to `_format_with_black` method in the `LocalizationKeyGenerator` class were commented out. The `_format_with_black` method in the `LocalizationKeyGenerator` class was also commented out. This method was responsible for formatting the Python file at the specified path using the Black code formatter.